backend/requirements.txt
# ===================================
# 后端依赖 - requirements.txt
# ===================================

# Web 框架
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# 数据库
sqlalchemy==2.0.30
alembic==1.13.1
aiomysql==0.2.0
pymysql==1.1.1

# Redis
redis==5.0.4
aioredis==2.0.1

# 认证 & 安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# AI & LLM
openai==1.30.0
anthropic==0.26.0
langchain==0.2.1
langchain-openai==0.1.8
langchain-anthropic==0.1.13
langchain-community==0.2.1

# 向量数据库（Qdrant）
qdrant-client==1.12.1
langchain-qdrant==0.1.2

# 文档处理
pypdf==4.2.0
python-docx==1.1.0
docx2txt==0.8
unstructured==0.18.21

# 工具
pydantic==2.7.1
pydantic-settings==2.3.0
httpx==0.27.0
tenacity==8.3.0
loguru==0.7.2

# 测试
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0
